Created transformer which removes ActiveRecord specific Arel from the AST
