+++ lib/arel/extensions/active_record_type_caster_map.rb
@@ -1,0 +1,7 @@
+module Arel
+  module Visitors
+    class Dot
+      alias visit_ActiveRecord_TypeCaster_Map terminal
+    end
+  end
+end
