+++ lib/arel/transformer.rb
@@ -1,4 +1,5 @@
 require_relative './transformer/add_schema_to_table'
+require_relative './transformer/remove_active_record_info'
 
 module Arel
   module Transformer
